backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fairlearn==0.10.0
aif360==0.6.0
shap==0.44.0
pandas==2.1.3
scikit-learn==1.3.2
reportlab==4.0.7
pytest==7.4.3
httpx==0.25.2
email-validator==2.1.0.post1
